fix: extract editor logic to `useEditor` hook and fix a few ui bugs
