+++ src/components/ui/Editor.tsx
@@ -1,188 +1,16 @@
 'use client';
 
+import useEditor from '@/hooks/use-editor';
 import TextareaAutosize from 'react-textarea-autosize';
-import { useForm } from 'react-hook-form';
-import { PostCreationRequest, PostValidator } from '@/validators/post';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { FC, useCallback, useEffect, useRef, useState } from 'react';
-import type EditorJS from '@editorjs/editorjs';
-import { uploadFiles } from '@/lib/uploadthing';
-import { toast } from '@/hooks/use-toast';
-import { useMutation } from '@tanstack/react-query';
-import axios from 'axios';
-import { usePathname, useRouter } from 'next/navigation';
 
 interface EditorProps {
   communityId: string;
   isModalOpen: boolean;
 }
 
-const Editor: FC<EditorProps> = ({ communityId, isModalOpen }) => {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<PostCreationRequest>({
-    resolver: zodResolver(PostValidator),
-    defaultValues: {
-      communityId,
-      title: '',
-      content: null,
-    },
-  });
-
-  const ref = useRef<EditorJS>();
-  const [isMounted, setIsMounted] = useState<boolean>(false);
-  const _titleRef = useRef<HTMLTextAreaElement>(null);
-  const pathname = usePathname();
-  const router = useRouter();
-
-  const initializeEditor = useCallback(async () => {
-    const EditorJS = (await import('@editorjs/editorjs')).default;
-    const Header = (await import('@editorjs/header')).default;
-    const Embed = (await import('@editorjs/embed')).default;
-    const Table = (await import('@editorjs/table')).default;
-    const List = (await import('@editorjs/list')).default;
-    const LinkTool = (await import('@editorjs/link')).default;
-    const InlineCode = (await import('@editorjs/inline-code')).default;
-    const ImageTool = (await import('@editorjs/image')).default;
-    const Code = (await import('@editorjs/code')).default;
-    // const Paragraph = (await import("@editorjs/paragraph")).default;
-
-    if (!ref.current) {
-      const editor = new EditorJS({
-        holder: 'editor',
-        onReady() {
-          ref.current = editor;
-        },
-        placeholder: "What's happening?",
-        inlineToolbar: true,
-        data: { blocks: [] },
-        tools: {
-          header: Header,
-          linkTool: {
-            class: LinkTool,
-            config: {
-              endpoint: '/api/link',
-            },
-          },
-          image: {
-            class: ImageTool,
-            config: {
-              uploader: {
-                async uploadByFile(file: File) {
-                  const [res] = await uploadFiles('imageUploader', {
-                    // Pass 'imageUploader' directly as the endpoint
-                    files: [file],
-                  });
-
-                  return {
-                    success: 1,
-                    file: {
-                      url: res.url,
-                    },
-                  };
-                },
-              },
-            },
-          },
-          list: List,
-          code: Code,
-          inlineCode: InlineCode,
-          table: Table,
-          embed: Embed,
-        },
-      });
-    }
-  }, []);
-
-  useEffect(() => {
-    if (typeof window !== 'undefined') {
-      setIsMounted(true);
-    }
-  }, []);
-
-  useEffect(() => {
-    if (Object.keys(errors).length) {
-      for (const [_key, value] of Object.entries(errors)) {
-        toast({
-          title: 'Action Failed',
-          description: (value as { message: string }).message,
-          variant: 'destructive',
-        });
-      }
-    }
-  }, [errors]);
-
-  useEffect(() => {
-    const init = async () => {
-      await initializeEditor();
-
-      setTimeout(() => {
-        _titleRef.current?.focus();
-      }, 0);
-    };
-
-    if (isMounted) {
-      init();
-
-      return () => {
-        ref.current?.destroy();
-        ref.current = undefined;
-      };
-    }
-  }, [isMounted, initializeEditor]);
-
-  const { mutate: createPost } = useMutation({
-    mutationFn: async ({
-      title,
-      content,
-      communityId,
-    }: PostCreationRequest) => {
-      const payload: PostCreationRequest = {
-        title,
-        content,
-        communityId,
-      };
-      const { data } = await axios.post('/api/community/post/create', payload);
-      return data;
-    },
-    onError: () => {
-      toast({
-        title: 'Action Failed',
-        description: 'Your post was not published, please try again later',
-        variant: 'destructive',
-      });
-    },
-
-    onSuccess: () => {
-      //z/community/submit => z/community
-
-      const newPathname = pathname.split('/').slice(0, -1).join('/');
-      router.push(newPathname);
-
-      router.refresh();
-
-      return toast({
-        description: 'Your post has been published.',
-        variant: 'default',
-      });
-    },
-  });
-
-  async function onSubmit(data: PostCreationRequest) {
-    const blocks = await ref.current?.save();
-
-    const payload: PostCreationRequest = {
-      title: data.title,
-      content: blocks,
-      communityId,
-    };
-
-    createPost(payload);
-  }
-
-  const { ref: titleRef, ...rest } = register('title');
+const Editor = ({ communityId, isModalOpen }: EditorProps) => {
+  const { onSubmit, handleSubmit, rest, titleRef, _titleRef } =
+    useEditor(communityId);
 
   return (
     <div className='w-full p-4 rounded-lg border-zinc-200 flex items-center justify-center'>
